+++ Controls/_gridNew/GridView.ts
@@ -2,8 +2,7 @@ import { ListView } from 'Controls/list';
 import { TemplateFunction } from 'UI/Base';
 import { TouchContextField as isTouch } from 'Controls/context';
 import { Logger} from 'UI/Utils';
-import { GridLayoutUtil } from 'Controls/grid';
-import { GridLadderUtil } from 'Controls/display';
+import { GridLadderUtil, GridLayoutUtil } from 'Controls/display';
 import * as GridTemplate from 'wml!Controls/_gridNew/Render/grid/GridView';
 import * as GridItem from 'wml!Controls/_gridNew/Render/grid/Item';
 import { prepareEmptyEditingColumns } from '../_grid/utils/GridEmptyTemplateUtil';
