+++ Controls/display.ts
@@ -82,6 +82,8 @@ import * as GridLadderUtil from './_display/utils/GridLadderUtil';
 export {GridLadderUtil};
 import isFullGridSupport from './_display/utils/GridSupportUtil';
 export {isFullGridSupport};
+import GridLayoutUtil from './_display/utils/GridLayoutUtil';
+export {GridLayoutUtil};
 import {default as GridCollection} from './_display/GridCollection';
 export {GridCollection};
 import {default as GridCollectionItem} from './_display/GridCollectionItem';
